# packages/shared/openapi.yaml
openapi: 3.1.0
info:
  title: DuSheng Panel API
  version: 0.1.0
  description: >
    REST contract for DuSheng Panel. The API manages users, device groups,
    nodes, tunnels, forwarding rules, protocol policies, install tokens,
    traffic samples, audit logs, and node-agent synchronization.
servers:
  - url: http://127.0.0.1:18888
    description: Local API server
  - url: https://panel.example.com
    description: Production reverse proxy
...
paths:
...
  /api/v1/forward-rules/batch/preview:
    post:
      tags: [User]
      operationId: previewForwardRuleBatch
      summary: Validate and allocate ports for a rule batch without committing
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ForwardRuleBatchWrite" } } }
      responses:
        "200":
          description: Batch preview.
          content: { application/json: { schema: { $ref: "#/components/schemas/ForwardRuleBatchResponse" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/forward-rules/batch:
    post:
      tags: [User]
      operationId: createForwardRuleBatch
      summary: Atomically create up to 100 forwarding rules
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ForwardRuleBatchWrite" } } }
      responses:
        "201":
          description: Batch created.
          content: { application/json: { schema: { $ref: "#/components/schemas/ForwardRuleBatchResponse" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access or refresh JWT.
    AgentBearerAuth:
      type: http
      scheme: bearer
      description: Node token returned by agent registration.
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    BaseModel:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ForwardRule:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - $ref: "#/components/schemas/ForwardRuleWrite"
        - type: object
          required: [status, inBytes, outBytes, revision, violationCount]
          properties:
            tenantId:
              type: [integer, "null"]
              minimum: 1
            status:
              $ref: "#/components/schemas/ForwardRuleStatus"
            inBytes:
              type: integer
              minimum: 0
              format: int64
            outBytes:
              type: integer
              minimum: 0
              format: int64
            revision:
              type: integer
              format: int64
            lastSyncError:
              type: string
            quotaSource:
              type: string
              enum: ["", user, tenant]
            violationCount:
              type: integer
              minimum: 0
    ForwardRuleWrite:
      type: object
      required: [tunnelId, name, protocol, remoteHost, remotePort]
      properties:
        userId:
          type: integer
          minimum: 1
          description: Required for admin callers; non-admin callers are forced to their own user ID.
        tunnelId:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 120
        protocol:
          $ref: "#/components/schemas/ForwardRuleProtocol"
        listenPort:
          type: integer
          minimum: 0
          maximum: 65535
          description: Zero or omitted asks the API to allocate a port from the tunnel entry group.
        remoteHost:
          type: string
          minLength: 1
          maxLength: 255
        remotePort:
          type: integer
          minimum: 1
          maximum: 65535
        strategy:
          type: string
          enum: [least_conn, round_robin, random, source_hash]
          default: least_conn
        status:
          type: string
          enum: [active, paused, disabled]
          description: active requests a resync; quota_exhausted is a system-managed status.
        protocolPolicyId:
          type: [integer, "null"]
          minimum: 1
          description: Rule-level override. Only administrators may set or change this field; other accounts inherit the existing rule, tunnel, or entry-group policy.
    ForwardRuleBatchWrite:
      type: object
      required: [rules]
      properties:
        rules:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: "#/components/schemas/ForwardRuleWrite" }
    ForwardRuleBatchResponse:
      type: object
      required: [items, count]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ForwardRule" }
        count: { type: integer, minimum: 0, maximum: 100 }
        preview: { type: boolean }
    ForwardRuleProtocol:
      type: string
      enum: [tcp, udp, tcp_udp]
    ForwardRuleStatus:
      type: string
      enum: [unsynced, syncing, active, paused, disabled, failed, protocol_violation, quota_exhausted]
...